Tidy up AssignmentStatement, ObjectPropertyAssignmentStatement
